test: Outsource common helpers to a shared module.
